The starting value of the stateid seqid field is defined
Elsewhere we claim it starts with 1.

(Alternatively: fix the other spot.  This fix makes us consistent with
rfc 5661; fixing the other spot would make us consistent with rfc 3530.

The mild inconsistency with rfc 3530 should be OK, I think, as long as
clients don't go so far as to error out if a server returns the wrong
initial seqid....)