use state-owner, open-owner, and lock-owner consistently.
I actually prefer the hyphenless versions, which are what we use on the
mailing lists and elsewhere (even in the protocol itself, as in
release_lockowner).

But we should be consistent in any case.  Especially eradicate a few
places where the term "lockowner" is used for the general (or even
specifically for the openowner) case.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>